feat: переработать архитектуру в части хранения фильмов, исправить неработающие тесты
